% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_registries.R
\name{get_all_registries}
\alias{get_all_registries}
\title{Retrieves the number of all registered events
(births, marriages, and deaths) according to the
arguments provided}
\usage{
get_all_registries(
  rd = NULL,
  year = "2020",
  month = "Todos",
  state = "Todos",
  wait = 10L
)
}
\arguments{
\item{rd}{An RSelenium webdriver.}

\item{year}{Year}

\item{month}{Month}

\item{state}{State}

\item{wait}{Waiting}
}
\value{
A data frame.
}
\description{
Retrieves the number of all registered events
(births, marriages, and deaths) according to the
arguments provided
}
