% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deaths.R
\name{get_deaths}
\alias{get_deaths}
\title{Retrieves the number of deaths from the specified
city or state.}
\usage{
get_deaths(
  rd = NULL,
  year = "2020",
  month = "Todos",
  state = "Todos",
  wait = 10L
)
}
\arguments{
\item{rd}{An RSelenium remote driver.}

\item{year}{An year.}

\item{month}{A month.}

\item{state}{A state.}

\item{wait}{Waiting time in seconds before attempting to
retrieve data.}
}
\value{
A data frame containing the required data.
}
\description{
Retrieves the number of deaths from the specified
city or state.
}
\examples{
\dontrun{
get_deaths(rd, "2015", "Junho")
}
}
