% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marriages.R
\name{get_marriages}
\alias{get_marriages}
\title{Retrieves the number of registered marriages
according to the arguments provided}
\usage{
get_marriages(
  rd = NULL,
  year = "2020",
  month = "Todos",
  state = "Todos",
  wait = 10L
)
}
\arguments{
\item{rd}{An RSelenium webdriver.}

\item{year}{Year}

\item{month}{Month}

\item{state}{State}

\item{wait}{Waiting}
}
\value{
A data frame.
}
\description{
Retrieves the number of registered marriages
according to the arguments provided
}
