% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{get_all_registries}
\alias{get_all_registries}
\title{Retrieves the number of registries in ARPEN website.}
\usage{
get_all_registries(
  rd = NULL,
  year = "2020",
  month = "Todos",
  state = "Todos",
  wait = 10L
)
}
\arguments{
\item{rd}{An initialized RSelenium webdriver.}

\item{year}{A numerical or character representing the year
you want to retrieve the registries from.}

\item{month}{A character representing the month name in Portuguese
or "Todos" for retrieving data from all the months in the
selected year.}

\item{state}{A character representing the State name in Portuguese or
"Todos" for retrieving data from all states.}

\item{wait}{An integer representing how long to wait before starting
the scraping operation after reaching
the Brazilian Association of Civil Registries website.}
}
\value{
A data frame containing the requested data.
}
\description{
Retrieves the number of the recorded vital events (all, births,
marriages, and deaths) available at the ARPEN website
\url{https://transparencia.registrocivil.org.br/registros}
according to the
arguments provided.

The functions work by navigating to the ARPEN website and
filling the form in order to generate a scrapable webpage.

The functions take a pre-initialized webDriver from the \link{RSelenium}
package (see example).
\itemize{
\item \code{get_all_registries()} Retrieves the number off all registries.
\item \code{get_births()} Retrieves the number of registered births.
\item \code{get_marriages()} Retrieves the number of registered marriages.
\item \code{get_deaths()} Retrieves the number of registered deaths.
}
}
\examples{
\dontrun{
rs <- RSelenium::rsDriver(browser = "firefox")
rd <- rs$client

get_all_registries(rd, "2018", "Janeiro", "Acre")
get_deaths(rd, "2020", "Maio", "Todos")

rd$quit()
rd$server$stop()
}
}
